niffler-e-2-e-tests-python/requirements.txt
selene>=2.0.0rc7
pytest
requests
pydantic
python-dotenv
faker

psycopg2-binary
sqlalchemy
sqlmodel
